chore: refactor worker into module
